Improve 3 functions, match 1
Widen four integer type declarations that caused MWCC to emit width and
signedness conversions retail does not have.

gm_801B5324 arg1 -> s32 (matches), fn_8016B388 arg1 -> int, mn_80230D18
arg2 -> int, and ifStock_802F8298_get_flag returns u8.

+++ src/melee/gm/gmallstar.c
@@ -393,7 +393,7 @@ static inline void gm_801B5324_inline(s8* char_ids, gm_803DEBE8_t* opp_data,
     }
 }
 
-void gm_801B5324(UnkAllstarData* arg0, u8 arg1)
+void gm_801B5324(UnkAllstarData* arg0, s32 arg1)
 {
     s8 chars[3];
     u8 colors[3];
